refactor: searchPath 중복 제거
PieceView에 Empty 추가
