+++ src/main/java/domain/piece/Piece.java
@@ -2,9 +2,13 @@ package domain.piece;
 
 import domain.Location;
 import domain.type.Color;
+import domain.type.Direction;
 import domain.type.PieceType;
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import view.PieceView;
 
 public abstract class Piece {
 
@@ -21,7 +25,22 @@ public abstract class Piece {
         return new EmptyPiece();
     }
 
-    abstract public List<Location> searchPath(final Location start, final Location end);
+    abstract protected boolean isNotMovable(final Location start, final Location end);
+    
+    public List<Location> searchPath(final Location start, final Location end) {
+        final Direction direction = Direction.find(start, end);
+        if (direction.equals(Direction.ELSE) || isNotMovable(start, end)) {
+            throw new IllegalArgumentException(PieceView.findSign(this) + IMPOSSIBLE_MOVE_ERROR_MESSAGE);
+        }
+        final int totalMove = Math.max(
+            Math.abs(start.getCol() - end.getCol()), Math.abs(start.getRow() - end.getRow()));
+        return IntStream.range(1, totalMove + 1)
+            .mapToObj(
+                count -> Location.of(
+                    start.getCol() + (direction.getColDiff() * count),
+                    start.getRow() + (direction.getRowDiff()) * count))
+            .collect(Collectors.toList());
+    }
 
     public boolean isSameType(final PieceType pieceType) {
         return this.pieceType.equals(pieceType);
